+++ src/components/feedback/index.tsx
@@ -1,0 +1,2 @@
+export { default as Feedback, feedbackTypeNames } from './Feedback';
+export { default as FeedbackForm } from './FeedbackForm';
